+++ R/01-plotFunctions.R
@@ -1,6 +1,6 @@
 makeMultiPlot <- function(valuesList, nMarginLines, combiType = "rowGrid", nGridCols = 2,
                           xAxisToHide = NULL, yAxisToHide = NULL, showSig = FALSE,
-                          referencePlot = NULL, sigLevel = 0.95
+                          referencePlot = NULL, sigLevel = 0.95, legendPosition
                           ){
 
   if (is.null(names(valuesList))) return(NULL)
@@ -39,6 +39,14 @@ makeMultiPlot <- function(valuesList, nMarginLines, combiType = "rowGrid", nGrid
                hideYAxis = (p %in% yAxisToHide),
                marginLine = 3 * (i - 1))
     }
+
+    # add legend to multiplot
+      addLegendServer(
+              id = "joinedPlot",
+              position = legendPosition,
+              title = "Color for data points",
+              legendEntries = names(valuesList),
+              fillColor = unlist(lapply(valuesList, function(x) x$plotStyle$dataPoints$color)))
   }
 
   if(combiType == "fullGrid") {
@@ -322,6 +330,8 @@ plotPredictions <- function(predData,
   uncertaintyLineType <- styleUncertainty$lineType
   uncertaintyWidth <- styleUncertainty$lineWidth
   uncertaintyHide <- styleUncertainty$hide
+  uncertaintyBandColor <- styleUncertainty$bandColor
+  uncertaintyBandOpacity <- styleUncertainty$bandOpacity
 
   plotRPred <- predData
 
@@ -353,22 +363,41 @@ plotPredictions <- function(predData,
     }
   }
 
+  upperValues <- getUncertaintyLimit(plotRPred,
+                                     type = errorType,
+                                     factor = uncertaintyFactor)$upper
+
+  lowerValues <- getUncertaintyLimit(plotRPred,
+                                     type = errorType,
+                                     factor = uncertaintyFactor)$lower
+
   plotLines(hide = uncertaintyHide,
-            getUncertaintyLimit(plotRPred,
-                                type = errorType,
-                                factor = uncertaintyFactor)$upper ~ plotRPred$xVar,
+            upperValues ~ plotRPred$xVar,
             lwd = uncertaintyWidth,
             lty = uncertaintyLineType,
             col = uncertaintyColor)
   plotLines(hide = uncertaintyHide,
-            getUncertaintyLimit(plotRPred,
-                                type = errorType,
-                                factor = uncertaintyFactor)$lower ~ plotRPred$xVar,
+            lowerValues ~ plotRPred$xVar,
             lwd = uncertaintyWidth,
             lty = uncertaintyLineType,
             col = uncertaintyColor)
+
+  # Add color band between model prediction uncertainty lines
+  polygon(c(plotRPred$xVar, rev(plotRPred$xVar), plotRPred$xVar[1]),
+          c(lowerValues, rev(upperValues), lowerValues[1]),
+          col = hex_with_opacity(hex = uncertaintyBandColor, opacity = uncertaintyBandOpacity))
+
   # lines((plotRPred$Estimation + 1.96 * plotRPred$SETOTAL) ~ plotRPred$xVar, lwd = 1, lty = 3)
   # lines((plotRPred$Estimation - 1.96 * plotRPred$SETOTAL) ~ plotRPred$xVar, lwd = 1, lty = 3)
+}
+
+#' Convert hex to hex with opacity
+#'
+#' @param hex color in hex format
+#' @param opacity opacity value between 0 and 1
+hex_with_opacity <- function(hex, opacity){
+  rgb_col <- col2rgb(hex)
+  rgb(rgb_col[1], rgb_col[2], rgb_col[3], opacity * 255, maxColorValue = 255)
 }
 
 plotLines <- function(hide, x, ...) {
@@ -411,4 +440,39 @@ cleanLabel <- function(dataColumns) {
 emptyPlot <- function(label = "No data available") {
   plot(0:1, 0:1, pch = NA, xlab = "", ylab = "")
   text(0.5, 0.5, label = label, col = "red")
+}
+
+
+
+#' selectInput to specify legend position
+#'
+#' @param id module id
+#' @param label label of button
+#' @param choices choices e.g. "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center"
+addLegendUI <- function(id, label = "Legend Position", choices = c("none", "topleft", "topright", "bottomright", "bottomleft")) {
+  ns <- NS(id)
+    selectizeInput(inputId = ns("legend"),
+                   label = label,
+                   choices = choices)
+}
+
+#' Add a legend to an existing plot
+#'
+#' @param id module id
+#' @param position legend position, e.g. "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center"
+#' @param title legend title
+#' @param legendEntries names of legend entries
+#' @param fillColor color for the legend symbols
+addLegendServer <- function(id, position, title, legendEntries, fillColor) {
+  moduleServer(
+    id,
+    function(input, output, session) {
+      if(position != "none"){
+      legend(x = position,
+             title = title,
+             legend = legendEntries,
+             fill = fillColor)
+      }
+    }
+  )
 }